# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Документация API вашего приложения",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Основной API сервер"
        }
    ],
    "paths": {
        "/api/activities": {
            "get": {
                "tags": [
                    "Деятельности"
                ],
                "summary": "Получить список всех видов деятельности",
                "operationId": "53e7a0b67f6097ed90245000764b2e5e",
                "responses": {
                    "200": {
                        "description": "Список всех видов деятельности",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Activity"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Деятельности"
                ],
                "summary": "Создать новую деятельность",
                "operationId": "8aa12114961e22767ecdad7c24be367e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Молочная продукция"
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "example": 5,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Деятельность создана"
                    },
                    "422": {
                        "description": "Нельзя создать деятельность глубже 3-го уровня"
                    }
                }
            }
        },
        "/api/activities/{id}": {
            "get": {
                "tags": [
                    "Деятельности"
                ],
                "summary": "Получить вид деятельности по ID",
                "operationId": "6d75cdd5535b9b12358448ec198faba4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID деятельности",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о деятельности"
                    },
                    "404": {
                        "description": "Деятельность не найдена"
                    }
                }
            },
            "put": {
                "tags": [
                    "Деятельности"
                ],
                "summary": "Обновить деятельность по ID",
                "operationId": "064379988a5fe2c974ae5ab18f0cddfd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID деятельности",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Еда и напитки"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Деятельность обновлена"
                    },
                    "404": {
                        "description": "Деятельность не найдена"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Деятельности"
                ],
                "summary": "Удалить деятельность по ID",
                "operationId": "d822403ad9a85feb951b1073858a072d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID деятельности",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Деятельность удалена"
                    },
                    "404": {
                        "description": "Деятельность не найдена"
                    }
                }
            }
        },
        "/api/activities/{id}/organizations": {
            "get": {
                "tags": [
                    "Деятельности"
                ],
                "summary": "Получить организации, связанные с данной деятельностью",
                "operationId": "78cac0f5a4f521f7269c4db31835c917",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID деятельности",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций, связанных с деятельностью"
                    },
                    "404": {
                        "description": "Деятельность не найдена"
                    }
                }
            }
        },
        "/api/activities/{id}/organizations-with-descendants": {
            "get": {
                "tags": [
                    "Деятельности"
                ],
                "summary": "Получить организации, связанные с данной деятельностью и её потомками",
                "operationId": "522653bce5c60cd51c2ca2f8a59b4e4e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID деятельности",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Организации, связанные с деятельностью и её потомками"
                    },
                    "404": {
                        "description": "Деятельность не найдена"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Activity": {
                "title": "Activity",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 5
                    },
                    "name": {
                        "type": "string",
                        "example": "Еда"
                    },
                    "parent_id": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "_lft": {
                        "type": "integer",
                        "example": 1
                    },
                    "_rgt": {
                        "type": "integer",
                        "example": 8
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}